Rename to follow conventions
Moved underdark to dimensions.underdark,. renamed  .dimension to
.dimensions.
