+++ src/main/java/wffirilat/betterghasts/dimensions/underdark/BlockUnderdarkPortal.java
@@ -1,4 +1,4 @@
-package wffirilat.betterghasts.dimension;
+package wffirilat.betterghasts.dimensions.underdark;
 
 import java.util.Random;
 
@@ -18,6 +18,7 @@ import net.minecraft.util.Direction;
 import net.minecraft.world.IBlockAccess;
 import net.minecraft.world.World;
 import wffirilat.betterghasts.blocks.ModBlocks;
+import wffirilat.betterghasts.dimensions.ModDimensions;
 import wffirilat.betterghasts.lib.Constants;
 import cpw.mods.fml.common.registry.GameRegistry;
 import cpw.mods.fml.relauncher.Side;
