+++ src/main/java/wffirilat/betterghasts/dimensions/underdark/WorldChunkManagerUnderdark.java
@@ -1,4 +1,4 @@
-package wffirilat.betterghasts.dimension;
+package wffirilat.betterghasts.dimensions.underdark;
 
 import java.util.Arrays;
 import java.util.List;
